@@ -3,7 +3,11 @@ package fr.maxlego08.menu.loader;
 import fr.maxlego08.menu.MenuItemStack;
 import fr.maxlego08.menu.api.InventoryManager;
 import fr.maxlego08.menu.exceptions.ItemEnchantException;
-import fr.maxlego08.menu.zcore.utils.*;
+import fr.maxlego08.menu.zcore.utils.Banner;
+import fr.maxlego08.menu.zcore.utils.Firework;
+import fr.maxlego08.menu.zcore.utils.LeatherArmor;
+import fr.maxlego08.menu.zcore.utils.Potion;
+import fr.maxlego08.menu.zcore.utils.ZUtils;
 import fr.maxlego08.menu.zcore.utils.loader.Loader;
 import org.bukkit.Color;
 import org.bukkit.DyeColor;
@@ -17,6 +21,8 @@ import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.ItemFlag;
 import org.bukkit.potion.PotionType;
 
+import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -41,24 +47,39 @@ public class MenuItemStackLoader extends ZUtils implements Loader<MenuItemStack>
      */
     public MenuItemStack load(YamlConfiguration configuration, String path, Object... objects) {
 
-        MenuItemStack menuItemStack = new MenuItemStack(this.manager);
+        File file = (File) objects[0];
+
+        MenuItemStack menuItemStack = new MenuItemStack(this.manager, file.getPath(), path);
         menuItemStack.setData(configuration.getInt(path + "data", 0));
         menuItemStack.setDurability(configuration.getInt(path + "durability", 0));
         menuItemStack.setAmount(configuration.getString(path + "amount", "1"));
         menuItemStack.setMaterial(configuration.getString(path + "material", null));
         menuItemStack.setUrl(configuration.getString(path + "url", null));
 
-        if (configuration.contains(path + "potion")) {
+        Color potionColor = getColor(configuration, path + "color", null);
 
+        try {
+            Material material = Material.valueOf(configuration.getString(path + "material", null));
+            String materialName = material.toString();
+            if (materialName.startsWith("LEATHER_")) {
+                Color armorColor = getColor(configuration, path + "color", Color.fromRGB(160, 101, 64));
+                String type = materialName.replace("LEATHER_", "");
+                menuItemStack.setLeatherArmor(new LeatherArmor(LeatherArmor.ArmorType.valueOf(type), armorColor));
+            }
+        } catch (Exception ignored) {
+        }
+
+        if (configuration.contains(path + "potion")) {
             PotionType type = PotionType.valueOf(configuration.getString(path + "potion", "REGEN").toUpperCase());
             int level = configuration.getInt(path + "level", 1);
             boolean splash = configuration.getBoolean(path + "splash", false);
             boolean extended = configuration.getBoolean(path + "extended", false);
 
             Potion potion = new Potion(type, level, splash, extended);
+            potion.setColor(potionColor);
             menuItemStack.setPotion(potion);
-
         }
+
 
         if (configuration.contains(path + "banner")) {
 
@@ -84,34 +105,8 @@ public class MenuItemStackLoader extends ZUtils implements Loader<MenuItemStack>
                 builder.trail(section.getBoolean("trail"));
                 builder.with(FireworkEffect.Type.valueOf(section.getString("type", "BALL")));
                 builder.withColor(getColors(section, "colors"));
-                builder.withFade(getColors(section, "fakeColors"));
+                builder.withFade(getColors(section, "fadeColors"));
                 menuItemStack.setFirework(new Firework(isStar, builder.build()));
-            }
-
-        }
-
-        // I don't know how to write here
-        if (configuration.contains(path + "color")) {
-
-            Material material = Material.valueOf(configuration.getString(path + "type", null));
-            String materialName = material.toString();
-
-            if (materialName.startsWith("LEATHER_")) {
-                String color = configuration.getString(path + "color", "");
-                Color armorColor;
-
-                String[] split = color.split(",");
-
-                if (split.length == 3) {
-                    armorColor = Color.fromRGB(Integer.parseInt(split[0]), Integer.parseInt(split[1]), Integer.parseInt(split[2]));
-                } else if (split.length == 4) {
-                    armorColor = Color.fromARGB(Integer.parseInt(split[0]), Integer.parseInt(split[1]), Integer.parseInt(split[2]), Integer.parseInt(split[3]));
-                } else {
-                    armorColor = Color.fromRGB(160, 101, 64);
-                }
-
-                String type = materialName.replace("LEATHER_", "");
-                menuItemStack.setLeatherArmor(new LeatherArmor(LeatherArmor.ArmorType.valueOf(type), armorColor));
             }
 
         }
@@ -122,7 +117,7 @@ public class MenuItemStackLoader extends ZUtils implements Loader<MenuItemStack>
         menuItemStack.setModelID(configuration.getInt(path + "modelID", 0));
 
         List<String> enchants = configuration.getStringList(path + "enchants");
-        Map<Enchantment, Integer> enchantments = new HashMap<Enchantment, Integer>();
+        Map<Enchantment, Integer> enchantments = new HashMap<>();
 
         for (String enchantString : enchants) {
 
@@ -153,13 +148,24 @@ public class MenuItemStackLoader extends ZUtils implements Loader<MenuItemStack>
             }
         }
 
-        List<ItemFlag> flags = configuration.getStringList(path + "flags").stream().map(this::getFlag)
-                .collect(Collectors.toList());
+        List<ItemFlag> flags = configuration.getStringList(path + "flags").stream().map(this::getFlag).collect(Collectors.toList());
 
         menuItemStack.setEnchantments(enchantments);
         menuItemStack.setFlags(flags);
 
         return menuItemStack;
+    }
+
+    private Color getColor(YamlConfiguration configuration, String key, Color def) {
+        String[] split = configuration.getString(key, "").split(",");
+
+        if (split.length == 3) {
+            return Color.fromRGB(Integer.parseInt(split[0]), Integer.parseInt(split[1]), Integer.parseInt(split[2]));
+        } else if (split.length == 4) {
+            return Color.fromARGB(Integer.parseInt(split[0]), Integer.parseInt(split[1]), Integer.parseInt(split[2]), Integer.parseInt(split[3]));
+        }
+
+        return def;
     }
 
     private List<Color> getColors(ConfigurationSection section, String key) {
@@ -181,8 +187,87 @@ public class MenuItemStackLoader extends ZUtils implements Loader<MenuItemStack>
     /**
      *
      */
-    public void save(MenuItemStack item, YamlConfiguration configuration, String path, Object... objects) {
+    public void save(MenuItemStack item, YamlConfiguration configuration, String path, File file, Object... objects) {
+        configuration.set(path + "data", item.getData());
+        configuration.set(path + "durability", item.getDurability());
+        configuration.set(path + "amount", item.getAmount());
+        configuration.set(path + "material", item.getMaterial());
+        configuration.set(path + "url", item.getUrl());
 
+        Potion potion = item.getPotion();
+        Firework firework = item.getFirework();
+        LeatherArmor leatherArmor = item.getLeatherArmor();
+        Banner banner = item.getBanner();
+
+        if (potion != null) {
+            Color c = potion.getColor();
+
+            configuration.set(path + "potion", potion.getType().toString());
+            configuration.set(path + "level", potion.getLevel());
+            configuration.set(path + "splash", potion.isSplash());
+            configuration.set(path + "extended", potion.hasExtendedDuration());
+
+            if (c != null)
+                configuration.set("color", c.getAlpha() + "," + c.getRed() + "," + c.getGreen() + "," + c.getBlue());
+        }
+
+        if (firework != null) {
+            ConfigurationSection fireworkSection = configuration.createSection(path + "firework");
+            FireworkEffect effect = firework.getEffect();
+            List<String> stringColors = new ArrayList<>();
+            effect.getColors().forEach(c -> stringColors.add(c.getAlpha() + "," + c.getRed() + "," + c.getGreen() + "," + c.getBlue()));
+            List<String> stringFadeColors = new ArrayList<>();
+            effect.getColors().forEach(c -> stringFadeColors.add(c.getAlpha() + "," + c.getRed() + "," + c.getGreen() + "," + c.getBlue()));
+
+            fireworkSection.set("star", firework.isStar());
+            fireworkSection.set("flicker", effect.hasFlicker());
+            fireworkSection.set("trail", effect.hasTrail());
+            fireworkSection.set("type", effect.getType().toString());
+
+
+            fireworkSection.set("colors", stringColors);
+            fireworkSection.set("fadeColors", stringFadeColors);
+        }
+
+        if (leatherArmor != null) {
+            Color c = leatherArmor.getColor();
+            if (c != null)
+                configuration.set("color", c.getAlpha() + "," + c.getRed() + "," + c.getGreen() + "," + c.getBlue());
+        }
+
+        if (banner != null) {
+            List<Pattern> patterns = banner.getPatterns();
+
+            configuration.set(path + "banner", banner.getBaseColor().toString());
+            if (patterns != null && !patterns.isEmpty()) {
+                List<String> stringPatterns = new ArrayList<>();
+                for (Pattern p : patterns) {
+                    stringPatterns.add(p.getColor() + ":" + p.getPattern());
+                }
+                configuration.set(path + "patterns", stringPatterns);
+            }
+        }
+
+        configuration.set(path + "lore", item.getLore());
+        configuration.set(path + "name", item.getDisplayName());
+        configuration.set(path + "glow", item.isGlowing());
+        configuration.set(path + "modelID", item.getModelID());
+
+        if (item.getEnchantments() != null && !item.getEnchantments().isEmpty()) {
+            List<String> stringEnchantments = item.getEnchantments().entrySet().stream().map(e -> e.getKey().toString() + "," + e.getValue().toString()).toList();
+
+            configuration.set(path + "enchants", stringEnchantments);
+        }
+
+        if (item.getFlags() != null && !item.getFlags().isEmpty()) {
+            configuration.set(path + "flags", item.getFlags().stream().map(Enum::toString).toList());
+        }
+
+        try {
+            configuration.save(file);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 
 }
