@@ -7,6 +7,9 @@ import fr.maxlego08.menu.exceptions.InventoryException;
 import fr.maxlego08.menu.zcore.utils.loader.Loader;
 import org.bukkit.configuration.file.YamlConfiguration;
 
+import java.io.File;
+import java.io.IOException;
+
 public class ActionPlayerDataLoader implements Loader<ActionPlayerData> {
 
     @Override
@@ -22,8 +25,18 @@ public class ActionPlayerDataLoader implements Loader<ActionPlayerData> {
     }
 
     @Override
-    public void save(ActionPlayerData object, YamlConfiguration configuration, String path, Object... objects) {
+    public void save(ActionPlayerData object, YamlConfiguration configuration, String path, File file, Object... objects) {
 
+        configuration.set(path + "type", "SET");
+        configuration.set(path + "key", object.getKey());
+        configuration.set(path + "value", object.getValue());
+        configuration.set(path + "seconds", object.getSeconds());
+
+        try {
+            configuration.save(file);
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 
 }
