@@ -1,9 +1,6 @@
 package fr.maxlego08.menu.zcore.enums;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 import org.bukkit.inventory.ItemStack;
 
@@ -67,6 +64,7 @@ public enum Message {
     DESCRIPTION_PLAYERS_KEYS("Returns the list of keys of a player"),
     DESCRIPTION_PLAYERS_CLEAR_ALL("Clear all player's data"),
     DESCRIPTION_PLAYERS_CLEAR_PLAYER("Clear player's data"),
+	DESCRIPTION_OPEN_MAIN_MENU("Open the main menu"),
 	
 	RELOAD("§aYou have just reloaded the configuration files."),
 	RELOAD_INVENTORY("§aYou have just reloaded the inventories files."),
@@ -114,7 +112,7 @@ public enum Message {
 	private List<String> messages;
 	private String message;
 	private Map<String, Object> titles = new HashMap<>();
-	private boolean use = true;
+	private boolean use;
 	private MessageType type = MessageType.TCHAT;
 
 	private ItemStack itemStack;
@@ -123,7 +121,7 @@ public enum Message {
 	 * 
 	 * @param message
 	 */
-	private Message(String message) {
+    Message(String message) {
 		this.message = message;
 		this.use = true;
 	}
@@ -136,7 +134,7 @@ public enum Message {
 	 * @param b
 	 * @param c
 	 */
-	private Message(String title, String subTitle, int a, int b, int c) {
+    Message(String title, String subTitle, int a, int b, int c) {
 		this.use = true;
 		this.titles.put("title", title);
 		this.titles.put("subtitle", subTitle);
@@ -151,7 +149,7 @@ public enum Message {
 	 * 
 	 * @param message
 	 */
-	private Message(String... message) {
+    Message(String... message) {
 		this.messages = Arrays.asList(message);
 		this.use = true;
 	}
@@ -160,7 +158,7 @@ public enum Message {
 	 * 
 	 * @param message
 	 */
-	private Message(MessageType type, String... message) {
+    Message(MessageType type, String... message) {
 		this.messages = Arrays.asList(message);
 		this.use = true;
 		this.type = type;
@@ -170,7 +168,7 @@ public enum Message {
 	 * 
 	 * @param message
 	 */
-	private Message(MessageType type, String message) {
+    Message(MessageType type, String message) {
 		this.message = message;
 		this.use = true;
 		this.type = type;
@@ -181,7 +179,7 @@ public enum Message {
 	 * @param message
 	 * @param use
 	 */
-	private Message(String message, boolean use) {
+    Message(String message, boolean use) {
 		this.message = message;
 		this.use = use;
 	}
@@ -207,7 +205,7 @@ public enum Message {
 	}
 
 	public List<String> getMessages() {
-		return messages == null ? Arrays.asList(message) : messages;
+		return messages == null ? Collections.singletonList(message) : messages;
 	}
 
 	public void setMessages(List<String> messages) {
