@@ -99,8 +99,8 @@ public enum DefaultFontInfo{
 	SPACE(' ', 3),
 	DEFAULT('a', 4);
 	
-	private char character;
-	private int length;
+	private final char character;
+	private final int length;
 	
 	DefaultFontInfo(char character, int length) {
 		this.character = character;
