@@ -9,6 +9,13 @@ import org.jetbrains.annotations.NotNull;
 public class Firework {
     private boolean isStar;
     private FireworkEffect effect;
+
+    /**
+     * Create a firework
+     *
+     * @param isStar use {@link Material#FIREWORK_STAR} or not
+     * @param effect the firework effect
+     */
     public Firework(boolean isStar, FireworkEffect effect) {
         this.isStar = isStar;
         this.effect = effect;
@@ -33,7 +40,7 @@ public class Firework {
     @NotNull
     public ItemStack toItemStack(int amount) {
         Material material = isStar ? Material.FIREWORK_STAR : Material.FIREWORK_ROCKET;
-        ItemStack itemStack = new ItemStack(material);
+        ItemStack itemStack = new ItemStack(material, amount);
         FireworkEffectMeta fem = (FireworkEffectMeta) itemStack.getItemMeta();
         fem.setEffect(effect);
         itemStack.setItemMeta(fem);
